Mini gets a favorite button!
I'm extremely fond of the Mini mode in Audiation. The only thing I felt it was lacking was a favorite button; of course this is a feature that needs to be implemented across the entire app - something I very well may do - but this just gives a sneak peek. Also increased the size of the window just a little bit because it looks more balanced with a bigger album art photo.